Moved helper functions to render progressive paginators to a common place
